fix(snapshot): stabilize pending pairing read
